+++ src/openvpn/mtcp.c
@@ -49,7 +49,7 @@ multi_create_instance_tcp(struct multi_context *m, struct link_socket *sock)
     {
         mi->real.proto = sock->info.proto;
         struct hash_element *he;
-        const uint32_t hv = hash_value(hash, &mi->real);
+        const uint64_t hv = hash_value(hash, &mi->real);
         struct hash_bucket *bucket = hash_bucket(hash, hv);
 
         multi_assign_peer_id(m, mi);
